feature(schema utils): Properly implement not following references; Schema utilities to dereference a Schema (#47)
* fix: add an example based on first item in list, if not null/empty

* fix: remove follow reference check; always use references in ObjVisitor

* fix: always create obj prop references for mapper and obj visitor

* chore: bump version, add JUnit platform suite dep.

* feature(schema utils): create utils and handlers for enums

The Schema class is a bit weird because it is typed but its enums are also typed.  These utilities and handlers are used to help sync them when unknown.  This helps to avoid unmatched types and avoid using generics/raws everywhere.

* feature(schema utils): create sorting util for least to most Schema refs

This is a topological sort determining which Schemas in a Map of schemas have zero/the least references to most references to other Schemas in the OpenAPI spec.

* feature(schema utils): utils for (deep) cloning complex Schemas

Schema does not have its own clone() method.  Schemas are fairly complex and have "hidden" structure, such as the allOf, anyOf, and oneOf items.  These are not apparent when printing out a Schema.  This cloning utility helps make sure that a new Schema has no unintended link/reference to internal data structures of another Schema.

* feature(schema utils): util for refs within a Schema, and methods to dereference a Schema

Methods to check for or collect references in a Schema.  Also, method for removing references from a Schema and placing the referenced Schema details in-line.

* fix: if FOLLOW_REFERENCES config flag is false, remove all Schema refs

* feature(schema utils): unit test helper to create test Schemas

* feature(schema utils): suite of classes related to Schema utils

* fix: remove redundant copying of list

* fix: support raw Schema typing for arg

* fix: only add ref classes, if FOLLOW_REFERENCES is true

* fix: revert back to allow unknown Schema types for arg

* fix: revert back using references all the time

* fix: add missing constant for config

* feature(schema utils): refactor and fix complex schema lists and enums; fix unit tests

* fix(schema utils): deref schemas if in config file; and also remove the unreferenced schemas

* refactor: move allowed class/download handling out of Mapper into YamlConfig

* fix: move dereferencing back to Mapper

* refactor: move applicable unit tests to YamlConfigTest

* refactor: move allowed class/download handling out of Mapper into YamlConfig

* feat(custom exit handler): create and use custom exit handler